+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Unreleased
 
+- Major: Add achievement diary notifications. (#67)
 - Major: Add combat achievement notifications. (#63)
 - Major: Add additional information for PvP deaths. (#55)
 - Major: Add boss kill count notifications. (#50)
+- Minor: Add note about Collection Log requiring in-game setting (#70)
 - Minor: Add config to ignore slayer tasks with low point values. (#61)
 - Bugfix: Prevent old clue or slayer data from entering later notifications. (#68)
+- Dev: Bump `runeliteVersion` to always use `latest.release` (#64)
 - Dev: Add plugin-hub icon. (#65)
 
 ## 1.0.3